Separate everything in site-creds form into different components and hooks
