+++ client/landing/stepper/declarative-flow/internals/steps-repository/site-migration-credentials/components/error-message.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+interface Props {
+	error?: {
+		message?: string;
+	};
+}
+
+export const ErrorMessage: React.FC< Props > = ( { error } ) => {
+	if ( ! error || ! error.message ) {
+		return null;
+	}
+
+	return <div className="site-migration-credentials__form-error">{ error.message }</div>;
+};
